Relax capabilities required for config access
The 'manage_options' capability is perhaps not the best choice here because read-only access to the config files does not actually entail managing any of the settings. There is no other capability that is entirely fitting.

The 'unfiltered_html' capability allows users with the 'editor' role to use HTML code including scripts. While this capability has nothing in principle to do with access to config files, it does imply a very high level of trust for users with that role. (Also, the 'unfiltered' part may very loosely fit for raw config file access even though there is no HTML involved.)

Allowing users with the 'editor' role config file read access seems much more in line with the way the SKGB uses Wordpress roles.

It may in theory be better to use a new capability with add_cap() rather than attaching new functionality to existing capabilities that technically weren't designed for it. However, add_cap() modifies the database, which also may make the plugin more brittle.